backend/requirements.txt
Flask
Flask-JWT-Extended
Flask-SQLAlchemy
Flask-Migrate
psycopg2-binary
Werkzeug
flask-cors
debugpy
python-dotenv==1.0.1
